Module cdp::liquidations

Constants

LIQ_QUEUE_REPLY_ID

STABILITY_POOL_REPLY_ID

USER_SP_REPAY_REPLY_ID

Functions

get_LP_pool_cAssets

If cAssets include an LP, remove the LP share denom and add its paired assets

liquidate

Confirms insolvency and calculates repayment amount, then sends liquidation messages to the modules if they have funds. If not, sell wall.

query_stability_pool_liquidatible

Returns leftover liquidatible amount from the stability pool

sell_wall

Returns router & lp withdraw messages for use in liquidations

sell_wall_using_ids

Uses Position info to create sell wall msgs