Module cdp::positions

Constants

BAD_DEBT_REPLY_ID

CLOSE_POSITION_REPLY_ID

WITHDRAW_REPLY_ID

Functions

check_for_empty_position

Checks if any cAsset amount is zero

close_position

Sell position collateral to fully repay debts. Max spread is used to ensure the full debt is repaid in lieu of slippage.

create_basket

Create the contract’s Basket. Validates params.

create_position

Create Position instance

credit_burn_msg

Creates a CosmosMsg to burn credit tokens

credit_mint_msg

Creates a CosmosMsg to mint tokens

deposit

Deposit collateral to existing position. New or existing collateral. Anyone can deposit, to any position. Owner restrictions for withdrawals.

edit_basket

Edit the contract’s Basket. Can’t edit basket id, current_position_id or credit_asset. Credit price can only be changed thru the accrue function. Validates parameters and updates the basket.

increase_debt

Increase debt of a position. Accrue and validate credit amount. Check for insolvency & update basket debt tally.

liq_repay

This is what the stability pool contract calls to repay for a liquidation and get its collateral distribution

mint_revenue

Mint Basket’s pending revenue to the specified address

read_price

Reads the price of an asset from storage

repay

Use credit to repay outstanding debt in a Position. Validates repayment & updates state. Note: Excess repayment defaults to the sending address.

store_price

Stores the price of an asset

withdraw

Withdraws assets from a position. Validates withdraw amount & updates state.