Module cdp::query

Functions

append_lp_price

Calculate LP share token value. Calculate LP price. Append price and value to lists.

calculate_avg_LTV

Calculations for avg_borrow_LTV, avg_max_LTV, total_value and cAsset_prices

get_asset_values

Calculate cAsset values & returns a tuple of (cAsset_values, cAsset_prices)

get_avg_LTV

Calculates the average LTV of a position. Returns avg_borrow_LTV, avg_max_LTV, total_value and cAsset_prices.

get_cAsset_ratios

Returns cAsset ratios & prices for a Position

insolvency_check

Uses a Position’s info to calculate if the user is insolvent. Returns insolvent, current_LTV and available fee.

insolvency_check_calc

Function handles calculations for the insolvency check

query_bad_debt

Returns Position info with bad debt in the Basket

query_basket_credit_interest

Returns Basket credit redemption interest rate

query_basket_debt_caps

query_basket_positions

Returns Positions in a Basket

query_collateral_rates

Returns cAsset interest rates for the Basket

query_position

Returns Position information

query_position_insolvency

Returns Position’s insolvency status

query_price

Function queries the price of an asset from the oracle. If the query is within the oracle_time_limit, it will use the stored price.

query_user_positions

Returns Positions for a given user