Module cdp::reply

Functions

handle_close_position_reply

On success, update position claims & attempt to withdraw leftover using a WithdrawMsg

handle_liq_queue_reply

Send the liquidation queue its collateral reward. If the SP wasn’t used, send leftovers to the sell wall.

handle_stability_pool_reply

The reply used to handle all liquidation leftovers. Prioritizes use of the SP, then the sell wall.

handle_user_sp_repay_reply

On error of an user’s Stability Pool repayment, leave leftover handling to the SP reply unless SP wasn’t called. If so, sell wall the leftover.

handle_withdraw_reply

Validate withdrawls by asserting that the amount withdrawn is less than or equal to the amount of the asset in the contract. Assert new cAssets amount was saved correctly.

sell_wall_in_reply

Builds sell wall & LP messages to add to list of messages