Module cdp::risk_engine

Functions

assert_basket_assets

Asserts that the assets provided are valid collateral assets in the basket

get_basket_debt_caps

Calculate the debt cap for each asset in the Basket using network liquidity

get_stability_pool_liquidity

Get total amount of debt token in the Stability Pool

update_basket_debt

Update the distribution of Basket debt per asset after Position debt increases

update_basket_tally

Update SupplyCap objects in Basket

update_debt_per_asset_in_position

Update the distribution of Basket debt per asset after Position collateral makeup changes